bruin-taxi/pipeline/assets/ingestion/requirements.txt
pandas==2.2.3
requests==2.32.3
pyarrow==18.1.0
python-dateutil==2.9.0.post0
